+++ docs/ReleaseNotes.md
@@ -1,3 +1,5 @@
 // Before any commit, use the command line: tsc --noEmit
 
-# 0.1.3.4 uuidv4 method moved to UtilService
+# 0.1.3.5 "allowMultipleDownloads" project rule
+
+- A method to change all ids from a projectDTO was added on Project Service.